przeniesienie autoryzacji na technologie jwt i kilka kosmetycznych zmian

+++ src/app/api/transaction/delete.ts
@@ -5,6 +5,7 @@ import {
   increaseWalletBalance,
 } from "@app/utils/db-actions/wallet";
 import { deleteTransaction } from "@app/utils/db-actions/transaction";
+import { verifySession } from "@app/utils/session";
 
 export async function deleteTransactionAPI(
   transactionId: number,
@@ -13,6 +14,7 @@ export async function deleteTransactionAPI(
   income: boolean,
   userId: number
 ) {
+  if (!(await verifySession()).isAuth) return null;
   await deleteTransaction(transactionId);
   if (income) decreaseWalletBalance(walletId, amount);
   else increaseWalletBalance(walletId, amount);
